seperate new viewmodel classes.
