+++ WalletWasabi.Fluent/Behaviors/ShowFlyoutOnPointerOverBehavior.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Reactive.Disposables;
 using System.Reactive.Linq;
 using Avalonia;
