+++ WalletWasabi.Fluent/ViewModels/Wallets/Send/PrivacyOptimisationLevel.cs
@@ -1,0 +1,7 @@
+namespace WalletWasabi.Fluent.ViewModels.Wallets.Send;
+
+public enum PrivacyOptimisationLevel
+{
+	Standard,
+	Better
+}
