+++ WalletWasabi.Fluent/ViewModels/Wallets/Send/PrivacySuggestionsFlyoutViewModel.cs
@@ -1,0 +1,27 @@
+using System.Collections.ObjectModel;
+using ReactiveUI;
+
+namespace WalletWasabi.Fluent.ViewModels.Wallets.Send;
+
+public partial class PrivacySuggestionsFlyoutViewModel : ViewModelBase
+{
+	[AutoNotify] private SuggestionViewModel? _previewSuggestion;
+	[AutoNotify] private SuggestionViewModel? _selectedSuggestion;
+	[AutoNotify] private bool _isOpen;
+
+	public PrivacySuggestionsFlyoutViewModel()
+	{
+		Suggestions = new ObservableCollection<SuggestionViewModel>();
+
+		this.WhenAnyValue(x => x.IsOpen)
+			.Subscribe(x =>
+			{
+				if (!x)
+				{
+					PreviewSuggestion = null;
+				}
+			});
+	}
+
+	public ObservableCollection<SuggestionViewModel> Suggestions { get; }
+}
