+++ WalletWasabi.Fluent/ViewModels/Wallets/Send/SuggestionViewModel.cs
@@ -1,0 +1,12 @@
+using WalletWasabi.Blockchain.TransactionBuilding;
+
+namespace WalletWasabi.Fluent.ViewModels.Wallets.Send;
+
+public class SuggestionViewModel : ViewModelBase
+{
+	public bool IsOriginal { get; protected set; }
+
+	public string Suggestion { get; set; }
+
+	public BuildTransactionResult TransactionResult { get; protected set; }
+}
